possibilité d'obtenir les permissions de chaque rôle

+++ erp_back/src/main/java/fr/iut_unilim/erp_back/dto/RoleResponse.java
@@ -1,3 +1,3 @@
 package fr.iut_unilim.erp_back.dto;
 
-public record RoleResponse(int id, String name) {}
+public record RoleResponse(long id, String name) {}
